- Move superuser create page to `superuser_create` module. - Use `CreateView` form class to automatically create the superuser
