+++ mathesar/users/password_reset.py
@@ -1,13 +1,9 @@
-from django.contrib.auth import get_user_model
-from django.contrib.auth.forms import SetPasswordForm, UserCreationForm
+from django.contrib.auth.forms import SetPasswordForm
 from django.contrib.auth.views import PasswordResetConfirmView
 from django.utils.decorators import method_decorator
 from django.views.decorators.cache import never_cache
 from django.views.decorators.debug import sensitive_post_parameters
 from django.utils.translation import gettext_lazy as _
-from django.views.generic import FormView
-
-User = get_user_model()
 
 
 class MathesarSetPasswordForm(SetPasswordForm):
@@ -40,22 +36,3 @@ class MathesarPasswordResetConfirmView(PasswordResetConfirmView):
     def form_valid(self, form):
         form.save()
         return super(PasswordResetConfirmView, self).form_valid(form)
-
-
-class SuperuserForm(UserCreationForm):
-    class Meta(UserCreationForm.Meta):
-        model = User
-        fields = UserCreationForm.Meta.fields
-
-    def save(self, commit=True):
-        user = super(SuperuserForm, self).save(commit=False)
-        user.is_superuser = True
-        if commit:
-            user.save()
-        return user
-
-
-class SuperuserFormView(FormView):
-    template_name = "registration/superuser_create.html"
-    form_class = SuperuserForm
-    success_url = "/auth/login"
